+++ src/main/java/com/nred/nuclearcraft/block/item/energy/EnergyItemBlock.java
@@ -2,7 +2,7 @@ package com.nred.nuclearcraft.block.item.energy;
 
 import com.nred.nuclearcraft.block.item.NCItemBlock;
 import com.nred.nuclearcraft.block_entity.internal.energy.EnergyConnection;
-import com.nred.nuclearcraft.item.energy.IChargeableItem;
+import com.nred.nuclearcraft.item.energy.IChargeableNBTItem;
 import com.nred.nuclearcraft.util.InfoHelper;
 import com.nred.nuclearcraft.util.UnitHelper;
 import net.minecraft.ChatFormatting;
@@ -18,7 +18,7 @@ import java.util.function.Supplier;
 
 import static com.nred.nuclearcraft.NuclearcraftNeohaul.MODID;
 
-public class EnergyItemBlock extends NCItemBlock implements IChargeableItem {
+public class EnergyItemBlock extends NCItemBlock implements IChargeableNBTItem {
     private final Supplier<Integer> capacity;
     private final Supplier<Integer> maxTransfer;
     private final EnergyConnection energyConnection;
@@ -32,23 +32,23 @@ public class EnergyItemBlock extends NCItemBlock implements IChargeableItem {
 
     @Override
     public void appendHoverText(ItemStack stack, TooltipContext context, List<Component> tooltip, TooltipFlag tooltipFlag) {
-        InfoHelper.infoLine(tooltip, ChatFormatting.LIGHT_PURPLE, Component.translatable(MODID + ".tooltip.energy_stored", UnitHelper.prefix(IChargeableItem.getEnergyStored(stack), getMaxEnergyStored(stack), 5, "FE")));
+        InfoHelper.infoLine(tooltip, ChatFormatting.LIGHT_PURPLE, Component.translatable(MODID + ".tooltip.energy_stored", UnitHelper.prefix(getEnergyStored(stack), getMaxEnergyStored(stack), 5, "FE")));
         super.appendHoverText(stack, context, tooltip, tooltipFlag);
     }
 
     @Override
     public boolean isBarVisible(ItemStack stack) {
-        return IChargeableItem.getEnergyStored(stack) > 0;
+        return getEnergyStored(stack) > 0;
     }
 
     @Override
     public int getBarColor(ItemStack stack) {
-        return FastColor.ARGB32.lerp((float) IChargeableItem.getEnergyStored(stack) / capacity.get(), ChatFormatting.RED.getColor(), ChatFormatting.GREEN.getColor());
+        return FastColor.ARGB32.lerp((float) getEnergyStored(stack) / capacity.get(), ChatFormatting.RED.getColor(), ChatFormatting.GREEN.getColor());
     }
 
     @Override
     public int getBarWidth(ItemStack stack) {
-        return Mth.ceil(Mth.clamp((double) IChargeableItem.getEnergyStored(stack) / capacity.get(), 0D, 1D) * 13f);
+        return Mth.ceil(Mth.clamp((double) getEnergyStored(stack) / capacity.get(), 0D, 1D) * 13f);
     }
 
     @Override