+++ src/main/java/com/nred/nuclearcraft/item/energy/EnergyItem.java
@@ -2,26 +2,27 @@ package com.nred.nuclearcraft.item.energy;
 
 import com.nred.nuclearcraft.NuclearcraftNeohaul;
 import com.nred.nuclearcraft.block_entity.internal.energy.EnergyConnection;
+import com.nred.nuclearcraft.item.TooltipItem;
+import com.nred.nuclearcraft.util.UnitHelper;
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
 import net.minecraft.util.FastColor;
-import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.function.Supplier;
 
-import static com.nred.nuclearcraft.helpers.SimpleHelper.getFEString;
 import static com.nred.nuclearcraft.registration.DataComponentRegistration.ENERGY_COMPONENT;
 
-public class EnergyItem extends Item implements IChargeableItem {
+public class EnergyItem extends TooltipItem implements IChargeableComponentItem {
     public final Supplier<Integer> capacity;
     public final Supplier<Integer> maxTransfer;
     private final EnergyConnection energyConnection;
 
-    public EnergyItem(Properties properties, Supplier<Integer> capacity, Supplier<Integer> maxTransfer, EnergyConnection energyConnection) {
-        super(properties);
+    public EnergyItem(Properties properties, Supplier<Integer> capacity, Supplier<Integer> maxTransfer, EnergyConnection energyConnection, String... tooltip) {
+        super(properties, Arrays.stream(tooltip).toList());
         this.capacity = capacity;
         this.maxTransfer = maxTransfer;
         this.energyConnection = energyConnection;
@@ -29,7 +30,7 @@ public class EnergyItem extends Item implements IChargeableItem {
 
     @Override
     public boolean isBarVisible(ItemStack stack) {
-        return stack.getOrDefault(ENERGY_COMPONENT, 0) > 0;
+        return getEnergyStored(stack) > 0;
     }
 
     @Override
@@ -39,12 +40,12 @@ public class EnergyItem extends Item implements IChargeableItem {
 
     @Override
     public int getBarWidth(ItemStack stack) {
-        return Math.round((float) stack.getOrDefault(ENERGY_COMPONENT, 0) / capacity.get() * 13.0F);
+        return Math.round((float) getEnergyStored(stack) / capacity.get() * 13.0F);
     }
 
     @Override
     public int getBarColor(ItemStack stack) {
-        return FastColor.ARGB32.lerp((float) stack.getOrDefault(ENERGY_COMPONENT, 0) / capacity.get(), ChatFormatting.RED.getColor(), ChatFormatting.GREEN.getColor());
+        return FastColor.ARGB32.lerp((float) getEnergyStored(stack) / capacity.get(), ChatFormatting.RED.getColor(), ChatFormatting.GREEN.getColor());
     }
 
     @Override
@@ -53,7 +54,8 @@ public class EnergyItem extends Item implements IChargeableItem {
             stack.set(ENERGY_COMPONENT, 0);
         }
 
-        tooltipComponents.add(Component.translatable(NuclearcraftNeohaul.MODID + ".tooltip.energy_stored", getFEString(stack.getOrDefault(ENERGY_COMPONENT, 0)), getFEString(capacity.get())).withStyle(ChatFormatting.LIGHT_PURPLE));
+        tooltipComponents.add(Component.translatable(NuclearcraftNeohaul.MODID + ".tooltip.energy_stored", UnitHelper.prefix(getEnergyStored(stack), getMaxEnergyStored(stack), 5, "FE")).withStyle(ChatFormatting.LIGHT_PURPLE));
+        super.appendHoverText(stack, context, tooltipComponents, tooltipFlag);
     }
 
     @Override