+++ src/main/java/com/nred/nuclearcraft/item/energy/IChargeableNBTItem.java
@@ -1,15 +1,14 @@
 package com.nred.nuclearcraft.item.energy;
 
-import com.nred.nuclearcraft.block_entity.internal.energy.EnergyConnection;
 import com.nred.nuclearcraft.block_entity.internal.energy.EnergyStorage;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.component.CustomData;
 
-public interface IChargeableItem {
-    static CompoundTag getEnergyStorageNBT(ItemStack stack) {
-        if (!(stack.getItem() instanceof IChargeableItem item)) {
+public interface IChargeableNBTItem extends IChargeableComponentItem {
+    default CompoundTag getEnergyStorageNBT(ItemStack stack) {
+        if (!(stack.getItem() instanceof IChargeableNBTItem item)) {
             return null;
         }
 
@@ -21,26 +20,16 @@ public interface IChargeableItem {
         return nbt.getCompound("energyStorage");
     }
 
-    static long getEnergyStored(ItemStack stack) {
+    default long getEnergyStored(ItemStack stack) {
         CompoundTag nbt = getEnergyStorageNBT(stack);
         return nbt == null ? 0L : nbt.getLong("energy") + nbt.getLong("waitingEnergy");
     }
 
-    static void setEnergyStored(ItemStack stack, long energy) {
+    default void setEnergyStored(ItemStack stack, long energy) {
         CompoundTag nbt = getEnergyStorageNBT(stack);
         if (nbt != null && nbt.contains("energy")) {
             nbt.putLong("energy", energy);
             stack.update(DataComponents.CUSTOM_DATA, CustomData.EMPTY, customData -> customData.update(tag -> tag.put("energyStorage", nbt)));
         }
     }
-
-    long getMaxEnergyStored(ItemStack stack);
-
-    int getMaxTransfer(ItemStack stack);
-
-    boolean canReceive(ItemStack stack);
-
-    boolean canExtract(ItemStack stack);
-
-    EnergyConnection getEnergyConnection(ItemStack stack);
 }