+++ Changelog.md
@@ -1,3 +1,2 @@
-- Updated to v2o.9.7
-- Fixed gas cooler and the pebble bed controller having the same recipe
-- Updated GuideME
+- Reworked energy items
+- Fixed multitool clearing too much